Use job_name instead of IP address in graphite metrics
Use a new property 'job_name' taken from components' registration
messages and pass it as a tag to handlers.  As soon as components
send this in their registration messages the graphite handler
will use the tag in the metric name instead of the IP address.

We do this because the metrics received from the bosh health monitor
also use the bosh job name (e.g. api_z0). These are a better means to
distinguish jobs with the same index in different AZs than IP
addresses.

Signed-off-by: Marco Voelz <marco.voelz@sap.com>